requirements.txt
Django>=4.0.0,<4.1
djangorestframework>=3.13.0,<3.14
psycopg2>=2.8.4,<3.0
pytest>=7.1.2,<8.0
pytest-django>=4.5.2,<5.0
pytest-cov>=3.0.0,<4.0
pytest-sugar>=0.9.4,<1.0
drf-spectacular>=0.22.0,<1.0
mixer>=7.2.0,<7.3.0
